\name{numeric or NULL-class}
\docType{class}
\alias{numeric or NULL-class}

\title{Class "numeric or NULL" }
\description{
  This class is a union that can represent either a numeric vector or a
  NULL value. 
}
\section{Objects from the Class}{A virtual Class: No objects may be
  created from it.}
\section{Methods}{
  No methods defined with class "numeric or NULL" in the signature.
}
\author{Kevin R. Coombes <krc@silicovore.com>}
\keyword{classes}
